Implement Covid Tweets API microservice

+++ covid_tweets/covid_tweets_api/src/main/java/com/dreamsoftware/covidtweets/persistence/model/UserEntity.java
@@ -17,7 +17,7 @@ import org.springframework.data.elasticsearch.annotations.FieldType;
 @NoArgsConstructor
 public class UserEntity {
 
-    @Field(type = FieldType.Long)
+    @Field(type = FieldType.Integer)
     private Long id;
 
     @Field(type = FieldType.Text)
